+++ base/test/put/put_by_part_test.dart
@@ -6,157 +6,15 @@ import 'package:qiniu_sdk_base/src/storage/methods/put/put_response.dart';
 import 'package:test/test.dart';
 import 'package:dio/adapter.dart';
 import 'package:dio/dio.dart';
-import 'package:dotenv/dotenv.dart' show env;
 import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
 
-import 'config.dart';
+import '../config.dart';
 import 'put_controller_builder.dart';
 
 void main() {
   configEnv();
-
-  final storage = Storage();
-
-  test('put should works well.', () async {
-    var pcb = PutControllerBuilder();
-    final response = await storage.putFile(
-      File('test_resource/test_for_put.txt'),
-      token,
-      options: PutOptions(
-        key: 'test_for_put.txt',
-        controller: pcb.putController,
-      ),
-    );
-    pcb.testAll();
-    expect(response.key, 'test_for_put.txt');
-
-    // 分片
-    pcb = PutControllerBuilder();
-    final file = File('test_resource/test_for_put_parts.mp4');
-    final putResponseByPart = await storage.putFile(
-      file,
-      token,
-      options: PutOptions(
-        key: 'test_for_put_parts.mp4',
-        partSize: 1,
-        controller: pcb.putController,
-      ),
-    );
-
-    expect(putResponseByPart, isA<PutResponse>());
-
-    pcb.testAll();
-  }, skip: !isSensitiveDataDefined);
-
-  test('put with returnBody should works well.', () async {
-    final auth = Auth(
-      accessKey: env['QINIU_DART_SDK_ACCESS_KEY'],
-      secretKey: env['QINIU_DART_SDK_SECRET_KEY'],
-    );
-
-    final token = auth.generateUploadToken(
-      putPolicy: PutPolicy(
-        insertOnly: 0,
-        returnBody: '{"ext": \$(ext)}',
-        scope: env['QINIU_DART_SDK_TOKEN_SCOPE'],
-        deadline: DateTime.now().millisecondsSinceEpoch + 3600,
-      ),
-    );
-
-    final response = await storage.putFile(
-      File('test_resource/test_for_put.txt'),
-      token,
-      options: PutOptions(key: 'test_for_put.txt'),
-    );
-
-    expect(response.rawData, {'ext': '.txt'});
-  }, skip: !isSensitiveDataDefined);
-
-  test('putFileBySingle should works well.', () async {
-    final file = File('test_resource/test_for_put.txt');
-    var pcb = PutControllerBuilder();
-    final response = await storage.putFileBySingle(
-      file,
-      token,
-      options: PutBySingleOptions(
-        key: 'test_for_put.txt',
-        controller: pcb.putController,
-      ),
-    );
-
-    pcb.testAll();
-    expect(response.key, 'test_for_put.txt');
-  }, skip: !isSensitiveDataDefined);
-
-  test('putFileBySingle can be cancelled.', () async {
-    final putController = PutController();
-    final key = 'test_for_put.txt';
-    final file = File('test_resource/test_for_put.txt');
-
-    final statusList = <RequestTaskStatus>[];
-    putController.addStatusListener((status) {
-      statusList.add(status);
-      if (status == RequestTaskStatus.Request) {
-        putController.cancel();
-      }
-    });
-    var future = storage.putFileBySingle(
-      file,
-      token,
-      options: PutBySingleOptions(key: key, controller: putController),
-    );
-    try {
-      await future;
-    } catch (error) {
-      expect(error, isA<StorageError>());
-      expect((error as StorageError).type, StorageErrorType.CANCEL);
-    }
-    expect(future, throwsA(TypeMatcher<StorageError>()));
-    expect(statusList[0], RequestTaskStatus.Init);
-    expect(statusList[1], RequestTaskStatus.Request);
-    expect(statusList[2], RequestTaskStatus.Cancel);
-
-    try {
-      // 预期同步发生
-      // ignore: unawaited_futures
-      storage.putFileBySingle(
-        file,
-        token,
-        options: PutBySingleOptions(key: key, controller: putController),
-      );
-    } catch (error) {
-      // 复用了相同的 controller，所以也会触发取消的错误
-      expect(error, isA<StorageError>());
-      expect((error as StorageError).type, StorageErrorType.CANCEL);
-    }
-
-    expect(future, throwsA(TypeMatcher<StorageError>()));
-
-    final response = await storage.putFileBySingle(
-      file,
-      token,
-      options: PutBySingleOptions(key: key),
-    );
-
-    expect(response, isA<PutResponse>());
-  }, skip: !isSensitiveDataDefined);
-
-  test('listenProgress on putFileBySingle method should works well.', () async {
-    final pcb = PutControllerBuilder();
-
-    final response = await storage.putFileBySingle(
-      File('test_resource/test_for_put.txt'),
-      token,
-      options: PutBySingleOptions(
-        key: 'test_for_put.txt',
-        controller: pcb.putController,
-      ),
-    );
-    expect(response.key, 'test_for_put.txt');
-    pcb.testAll();
-  }, skip: !isSensitiveDataDefined);
-
   test('putFileByPart should works well.', () async {
+    final storage = Storage();
     final pcb = PutControllerBuilder();
     var callnumber = 0;
     pcb.putController.addSendProgressListener((percent) {
@@ -227,7 +85,7 @@ void main() {
       expect((error as StorageError).type, StorageErrorType.CANCEL);
     }
     expect(future, throwsA(TypeMatcher<StorageError>()));
-    pcb.testStatus([
+    pcb.testStatus(targetStatusList: [
       RequestTaskStatus.Init,
       RequestTaskStatus.Request,
       RequestTaskStatus.Cancel
@@ -417,13 +275,9 @@ void main() {
     expect(cacheProvider.value.length, 0);
   }, skip: !isSensitiveDataDefined);
 
-  test('listenProgress on putFileByPart method should works well.', () async {
-    final putController = PutController();
-
-    double _percent;
-    putController.addSendProgressListener((percent) {
-      _percent = percent;
-    });
+  test('putFileByPart\'s status and progress should works well.', () async {
+    final storage = Storage();
+    final pcb = PutControllerBuilder();
 
     final response = await storage.putFileByPart(
       File('test_resource/test_for_put_parts.mp4'),
@@ -431,11 +285,13 @@ void main() {
       options: PutByPartOptions(
         key: 'test_for_put_parts.mp4',
         partSize: 1,
-        controller: putController,
+        controller: pcb.putController,
       ),
     );
     expect(response, isA<PutResponse>());
-    expect(_percent, equals(1));
+    pcb
+      ..testProcess()
+      ..testStatus(targetProgressList: [0.001, 0.01, 0.99, 1]);
   }, skip: !isSensitiveDataDefined);
 }
