+++ base/test/put/put_controller_builder.dart
@@ -4,12 +4,14 @@ import 'package:test/test.dart';
 class PutControllerBuilder {
   final putController = PutController();
   final statusList = <RequestTaskStatus>[];
+  final progressList = <double>[];
   double _sendPercent, _totalPercent;
 
   PutControllerBuilder() {
     putController
       ..addStatusListener(statusList.add)
       ..addProgressListener((percent) {
+        progressList.add(percent);
         _totalPercent = percent;
       })
       ..addSendProgressListener((percent) {
@@ -30,12 +32,17 @@ class PutControllerBuilder {
   }
 
   // 任务执行完成后执行此方法
-  void testStatus([List<RequestTaskStatus> targetStatusList]) {
+  void testStatus({
+    List<RequestTaskStatus> targetStatusList,
+    List<double> targetProgressList,
+  }) {
     targetStatusList ??= [
       RequestTaskStatus.Init,
       RequestTaskStatus.Request,
       RequestTaskStatus.Success
     ];
+    targetProgressList ??= [0.001, 0.99, 1];
     expect(statusList, equals(targetStatusList));
+    expect(progressList, containsAllInOrder(targetProgressList));
   }
 }
