+++ base/test/put/put_controller_test.dart
@@ -2,7 +2,7 @@ import 'package:qiniu_sdk_base/src/storage/storage.dart';
 import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
 import 'package:test/test.dart';
 
-import 'config.dart';
+import '../config.dart';
 
 void main() {
   configEnv();
